style(ui): Update course laboratories table styles to match the participants table
